Update
Negating the boolean table of correlated values within the algorithm
caused it to take too much time. Instead I went back to my original idea
that is negating each boolean value when it is referred to. This change
dropped the algorithm time from 93ms to 27ms on a dataset with 99
attributes.