Update 7-31-17
- Added new datasets, removed old ones for lack of original source
- Also moved datasets into a string array rather than separate strings
as it is easier to call Arrays.asList() rather than update every class
that calls a single dataset from the datasets class
- Added an icon to accompany the GUI
- When the user clicks on a node or a bar in the bar graph or line
graph, the bar/node will pop out a stackpane that shows the bar/node
value.
- When using the GUI the Multiple Regression process is done through a
thread as to not crash the GUI and if there are any problems with the
GUI becoming unresponsive affecting the process, this should hopefully
fix it.